requirements.txt
crewai==1.9.3
crewai-tools==0.38.1
streamlit==1.41.0
groq==0.13.0
supabase==2.10.0
sentence-transformers==3.3.1
chromadb==0.5.23
PyMuPDF==1.25.1
python-docx==1.1.2
pandas==2.2.3
plotly==5.24.1
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.10.4
tiktoken==0.8.0
